% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mixmodel.R
\name{cv.mix.model}
\alias{cv.mix.model}
\alias{print.cv.mix.model}
\alias{plot.cv.mix.model}
\title{#' A function to estimate alp and F.s given a data vector via Cross validation}
\usage{
cv.mix.model(
  data,
  folds = 10,
  reps = 1,
  cn.s = NULL,
  cn.length = NULL,
  gridsize = 200
)

\method{print}{cv.mix.model}(x, ...)

\method{plot}{cv.mix.model}(x, ...)
}
\arguments{
\item{data}{the numeric vector. Ideally data from a two component mixture model.}

\item{folds}{Number of folds used for cross-validation, default is 10.}

\item{reps}{Number of replications for cross-validation, default is 1.}

\item{cn.s}{A vector of numbers to be used for cross-validation, a vector of values. default is equally spaced grid of 100 values between `.001 * log(log(n))` to `.2* log(log(n))`, where n is length(data).}

\item{cn.length}{A number of equally spaced tuning parameter (between `.001 * log(log(n))` and `.2* log(log(n))`) values to search from, default is 100.}

\item{gridsize}{An integer. Number of equally spaced points (between 0 and 1) to evaluate the distance function. Larger values are more computationally intensive but also lead to more accurate estimates, default is 600.}
}
\value{
An object of class `cv.mix.model`, a list including the elements:
\item{alp.hat}{estimate of alpha.}
#' \item{cn.cv}{value of the chosen tuning parameter via cross validation for the final estimate of alpha. }
\item{Fs.hat}{A list containing with elements `x` and `y` values for the function estimate of `F.s`}
\item{dist.out}{An object of the class `dist.fun` using the complete data.gen}
\item{score}{ vector of cross validation scores at the input values of `cn.s`.}
}
\description{
#' A function to estimate alp and F.s given a data vector via Cross validation
}
\examples{
data.gen<- function( n, alpha){
  ind<- rbinom(n, 1, alpha)
  data<- ind* rbeta(n, 1,5) + (1-ind)* runif(n,0,1)
  return(data)
}
data.1<- data.gen(n = 200, alpha = .1)
est.cv <- cv.mix.model(data.1, folds = 10, reps = 1, cn.length =20 , gridsize = 200)
}
