% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_mixmodel.R
\name{cv.score}
\alias{cv.score}
\title{Computes the cross validataion score at c.n. Function used inside `cv.mix.model`}
\usage{
cv.score(tr.data, test.data, c.n)
}
\arguments{
\item{tr.data}{Training data vector}

\item{test.data}{Testing data vector}

\item{c.n}{a positive number, default value is `.1 log(log(n)`, where `n` is the length of the vector.}
}
\value{
Numeric value of the cross validation score at the input
}
\description{
Computes the cross validataion score at c.n. Function used inside `cv.mix.model`
}
