% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_est.R
\name{den.mix.model}
\alias{den.mix.model}
\title{A function to compute the estimate of f.s (the density corresponding to F.s) when f.s is known to be either decreasing or increasing.}
\usage{
den.mix.model(input, dec.density = TRUE)
}
\arguments{
\item{input}{An element of class `cv.mix.model` or `mix.model`}

\item{dec.density}{A logical variable. If FALSE then density assumed to be increasing. Default is TRUE.}
}
\value{
A piecewise function as a list.
\item{x}{data points where the piecewise density has changes.}
\item{y}{value of density at the above points}
}
\description{
A function to compute the estimate of f.s (the density corresponding to F.s) when f.s is known to be either decreasing or increasing.
}
\examples{
data.gen<- function( n, alpha){
	ind<- rbinom(n, 1, alpha)
	data<- ind* rbeta(n, 1,5) + (1-ind)* runif(n,0,1)
	return(data)
 }
 data.1<- data.gen(n = 1000, alp = .2)
est.fixed <- mix.model(data.1, method = "fixed"
, c.n = .05*log(log(length(data.1))), gridsize = 600)
out <- den.mix.model(est.fixed)
plot(est.fixed$Fs.hat, type= "l", xlab= "x", ylab= "F_s")
plot(out, type="l", xlab="x", ylab= "f_s")
}
