% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_calc.R
\name{dist.calc}
\alias{dist.calc}
\alias{print.dist.calc}
\alias{plot.dist.calc}
\title{Title}
\usage{
dist.calc(data, gridsize = 200)

\method{print}{dist.calc}(x, ...)

\method{plot}{dist.calc}(x, ...)
}
\arguments{
\item{data}{the numeric vector. Ideally data from a two component mixture model.}

\item{gridsize}{An integer. Number of equally spaced points (between 0 and 1) to evaluate the distance function. Larger values are more computationally intensive but also lead to more accurate estimates, default is 600.}
}
\value{
An object of class `dist.calc`, a list including the elements:
\item{distance}{A vector of the distance computed at a grid of alpha as created by `gridsize`.}
\item{grid.pts}{A vector of equally spaced points between 0 and 1 of length `gridsize`}
}
\description{
Title
}
