% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_mix_model.R
\name{mix.model}
\alias{mix.model}
\alias{print.mix.model}
\alias{plot.mix.model}
\title{A function to estimate  mixing proportion and F.s (the cdf for the unknown component) given a data vector from a two component mixture model. This also provides inference for the mixing proportion.}
\usage{
mix.model(
  data,
  method = c("lwr.bnd", "fixed", "cv"),
  c.n = NULL,
  folds = 10,
  reps = 1,
  cn.s = NULL,
  cn.length = 100,
  gridsize = 600
)

\method{print}{mix.model}(x, ...)

\method{plot}{mix.model}(x, ...)
}
\arguments{
\item{data}{the numeric vector. Ideally data from a two component mixture model.}

\item{method}{A  string indicating which method to use and what is the goal. (a) `fixed` : it will compute the estimate based on the value of c.n; (b) `cv' : use cross-validation for choosing c.n (tuning parameter); (c) `lwr.bnd` : computes 95\% lower confidence bound for the mixing proportion.}

\item{c.n}{a positive number, default value is `.1 log(log(n)`, where `n` is the length of the vector.}

\item{folds}{Number of folds used for cross-validation, default is 10.}

\item{reps}{Number of replications for cross-validation, default is 1.}

\item{cn.s}{A vector of numbers to be used for cross-validation, a vector of values. default is equally spaced grid of 100 values between `.001 * log(log(n))` to `.2* log(log(n))`, where n is length(data).}

\item{cn.length}{A number of equally spaced tuning parameter (between `.001 * log(log(n))` and `.2* log(log(n))`) values to search from, default is 100.}

\item{gridsize}{An integer. Number of equally spaced points (between 0 and 1) to evaluate the distance function. Larger values are more computationally intensive but also lead to more accurate estimates, default is 600.}
}
\value{
An object of class `mix.model`, a list including the elements:
\item{alp.hat}{estimate of alpha. NULL if `method' is `lwr.bnd'.}
\item{alp.Lwr}{95\% lower confidence bound of alpha. NULL if `method` is `cv` or `fixed`.}
\item{Fs.hat}{A list containing with elements `x` and `y` values for the function estimate of `F.s`}
\item{dist.out}{An object of the class `dist.calc` using the complete data.gen}
\item{c.n}{ Value of the tuning parameter used to compute the final estimate.}
\item{cv.out}{An object of class `cv.mix.model`. The object is NULL if method is `fixed`.}
}
\description{
A function to estimate  mixing proportion and F.s (the cdf for the unknown component) given a data vector from a two component mixture model. This also provides inference for the mixing proportion.
}
\examples{
data.gen<- function( n, alpha){
  ind<- rbinom(n, 1, alpha)
  data<- ind* rbeta(n, 1,5) + (1-ind)* runif(n,0,1)
  return(data)
}
data.1<- data.gen(n = 5000, alpha = .1)
est.lwr.bnd <- mix.model(data.1, method = "lwr.bnd", gridsize = 600)
}
